+++ components/common/HapticTab.tsx
