+++ components/common/TabBarBackground.tsx
