+++ components/common/ThemedSafeView.tsx
@@ -1,0 +1,24 @@
+import { SafeAreaView, View, type ViewProps } from "react-native";
+
+import { useThemeColor } from "@/hooks/useThemeColor";
+
+export type ThemedViewProps = ViewProps & {
+    lightColor?: string;
+    darkColor?: string;
+};
+
+export function ThemedView({
+    style,
+    lightColor,
+    darkColor,
+    ...otherProps
+}: ThemedViewProps) {
+    const backgroundColor = useThemeColor(
+        { light: lightColor, dark: darkColor },
+        "background"
+    );
+
+    return (
+        <SafeAreaView style={[{ backgroundColor }, style]} {...otherProps} />
+    );
+}
